+++ libs/data-access-rooms/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/rooms.service';
+export * from './lib/room-settings.service';
 export * from './lib/spaces.service';
 export * from './lib/unread-aggregator.service';